+++ trainer/graph_pair_trainer.py
@@ -388,7 +388,7 @@ class GraphPairTrainer(BaseTrainer):
             'loss': loss,
             'boxLoss': boxLoss,
             'relLoss': relLoss,
-            'edgePredLens':np.array([numEdgePred,numBoxPred,numEdgePred+numBoxPred,-1],dtype=np.float),
+            'edgePredLens':np.array([numEdgePred,numBoxPred,numEdgePred+numBoxPred,-1],dtype=np.float32),
             'rel_recall':eRecall,
             #'rel_prec': ePrec,
             'rel_fullPrec':fullPrec,
